# backend/docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "NeutriAI API",
    "description": "AI nutrition, fitness and lifestyle platform. All endpoints except `/healthz`, `/v1/billing/plans` and `/v1/webhooks/stripe` require a Supabase bearer token.",
    "version": "1.0.0"
  },
  "paths": {
...
    "/v1/recipes": {
      "post": {
        "tags": [
          "recipes"
        ],
        "summary": "Create Recipe",
        "operationId": "create_recipe_v1_recipes_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RecipeIn"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecipeOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "recipes"
        ],
        "summary": "Browse",
        "operationId": "browse_v1_recipes_get",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Q"
            }
          },
          {
            "name": "tag",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Tag"
            }
          },
          {
            "name": "cuisine",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cuisine"
            }
          },
          {
            "name": "max_kcal",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Max Kcal"
            }
          },
          {
            "name": "mine",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false,
              "title": "Mine"
            }
          },
          {
            "name": "saved",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false,
              "title": "Saved"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 50,
              "default": 20,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 0,
              "title": "Offset"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RecipeOut"
                  },
                  "title": "Response Browse V1 Recipes Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/recipes/{recipe_id}": {
      "get": {
        "tags": [
          "recipes"
        ],
        "summary": "Get Recipe",
        "operationId": "get_recipe_v1_recipes__recipe_id__get",
        "parameters": [
          {
            "name": "recipe_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Recipe Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecipeOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "recipes"
        ],
        "summary": "Delete Recipe",
        "operationId": "delete_recipe_v1_recipes__recipe_id__delete",
        "parameters": [
          {
            "name": "recipe_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Recipe Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Ok"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/recipes/{recipe_id}/save": {
      "post": {
        "tags": [
          "recipes"
        ],
        "summary": "Save Recipe",
        "operationId": "save_recipe_v1_recipes__recipe_id__save_post",
        "parameters": [
          {
            "name": "recipe_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Recipe Id"
            }
          },
          {
            "name": "folder",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "default",
              "title": "Folder"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Ok"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "recipes"
        ],
        "summary": "Unsave Recipe",
        "operationId": "unsave_recipe_v1_recipes__recipe_id__save_delete",
        "parameters": [
          {
            "name": "recipe_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Recipe Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Ok"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/recipes/{recipe_id}/adapt": {
      "post": {
        "tags": [
          "recipes"
        ],
        "summary": "Adapt Recipe",
        "description": "Rewrite a recipe for this user's allergies, diet, macros and schedule.",
        "operationId": "adapt_recipe_v1_recipes__recipe_id__adapt_post",
        "parameters": [
          {
            "name": "recipe_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Recipe Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AdaptRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdaptedRecipe"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AdaptRequest": {
        "properties": {
          "target_servings": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 50.0,
                "minimum": 1.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Target Servings"
          },
          "max_kcal_per_serving": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 3000.0,
                "minimum": 100.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Kcal Per Serving"
          },
          "max_carbs_g": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Carbs G"
          },
          "min_protein_g": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Min Protein G"
          },
          "honor_allergies": {
            "type": "boolean",
            "title": "Honor Allergies",
            "default": true
          },
          "honor_diet_mode": {
            "type": "boolean",
            "title": "Honor Diet Mode",
            "default": true
          },
          "consider_fasting_window": {
            "type": "boolean",
            "title": "Consider Fasting Window",
            "default": true
          },
          "consider_workout_load": {
            "type": "boolean",
            "title": "Consider Workout Load",
            "default": true
          },
          "extra_notes": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 400
              },
              {
                "type": "null"
              }
            ],
            "title": "Extra Notes"
          },
          "save_as_fork": {
            "type": "boolean",
            "title": "Save As Fork",
            "default": true
          }
        },
        "type": "object",
        "title": "AdaptRequest",
        "description": "Ask the AI to rewrite a recipe for this user's constraints."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "IngredientIn": {
        "properties": {
          "raw_text": {
            "type": "string",
            "maxLength": 200,
            "minLength": 1,
            "title": "Raw Text"
          },
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "quantity": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Quantity"
          },
          "unit": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Unit"
          },
          "grams": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Grams"
          },
          "is_optional": {
            "type": "boolean",
            "title": "Is Optional",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "raw_text"
        ],
        "title": "IngredientIn"
      },
...
      "Ok": {
        "properties": {
          "ok": {
            "type": "boolean",
            "title": "Ok",
            "default": true
          },
          "message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Message"
          }
        },
        "type": "object",
        "title": "Ok"
      },
...
      "RecipeIn": {
        "properties": {
          "title": {
            "type": "string",
            "maxLength": 120,
            "minLength": 2,
            "title": "Title"
          },
          "summary": {
            "type": "string",
            "title": "Summary",
            "default": ""
          },
          "photo_paths": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Photo Paths",
            "default": []
          },
          "categories": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Categories",
            "default": []
          },
          "cuisine": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cuisine"
          },
          "servings": {
            "type": "integer",
            "maximum": 50.0,
            "minimum": 1.0,
            "title": "Servings",
            "default": 1
          },
          "prep_minutes": {
            "type": "integer",
            "title": "Prep Minutes",
            "default": 0
          },
          "cook_minutes": {
            "type": "integer",
            "title": "Cook Minutes",
            "default": 0
          },
          "difficulty": {
            "type": "integer",
            "maximum": 5.0,
            "minimum": 1.0,
            "title": "Difficulty",
            "default": 2
          },
          "tags": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Tags",
            "default": []
          },
          "is_public": {
            "type": "boolean",
            "title": "Is Public",
            "default": true
          },
          "ingredients": {
            "items": {
              "$ref": "#/components/schemas/IngredientIn"
            },
            "type": "array",
            "minItems": 1,
            "title": "Ingredients"
          },
          "steps": {
            "items": {
              "$ref": "#/components/schemas/StepIn"
            },
            "type": "array",
            "minItems": 1,
            "title": "Steps"
          }
        },
        "type": "object",
        "required": [
          "title",
          "ingredients",
          "steps"
        ],
        "title": "RecipeIn"
      },
      "RecipeOut": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "author_id": {
            "type": "string",
            "title": "Author Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "summary": {
            "type": "string",
            "title": "Summary",
            "default": ""
          },
          "photo_paths": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Photo Paths",
            "default": []
          },
          "cuisine": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cuisine"
          },
          "servings": {
            "type": "integer",
            "title": "Servings"
          },
          "prep_minutes": {
            "type": "integer",
            "title": "Prep Minutes"
          },
          "cook_minutes": {
            "type": "integer",
            "title": "Cook Minutes"
          },
          "difficulty": {
            "type": "integer",
            "title": "Difficulty"
          },
          "tags": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Tags",
            "default": []
          },
          "steps": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Steps",
            "default": []
          },
          "ingredients": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Ingredients",
            "default": []
          },
          "per_serving": {
            "$ref": "#/components/schemas/Macros",
            "default": {
              "kcal": 0.0,
              "protein_g": 0.0,
              "carbs_g": 0.0,
              "fat_g": 0.0,
              "fiber_g": 0.0,
              "sugar_g": 0.0,
              "sodium_mg": 0.0
            }
          },
          "is_ai_generated": {
            "type": "boolean",
            "title": "Is Ai Generated",
            "default": false
          },
          "adaptation_note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Adaptation Note"
          },
          "like_count": {
            "type": "integer",
            "title": "Like Count",
            "default": 0
          },
          "save_count": {
            "type": "integer",
            "title": "Save Count",
            "default": 0
          }
        },
        "type": "object",
        "required": [
          "id",
          "author_id",
          "title",
          "servings",
          "prep_minutes",
          "cook_minutes",
          "difficulty"
        ],
        "title": "RecipeOut"
      },
...
    }
  }
}